+++ Chess/source/ChessThings/Board.h
@@ -12,7 +12,7 @@ struct canCastle
 class Board
 {
 private:
-	std::array<unsigned int, 64> BoardSquare;
+	std::array<uint8_t, 64> BoardSquare;
 	const std::string FEN;
 	size_t IndexOfSideToMove;//before side to move
 	size_t IndexOfCastling;//before castling ability
@@ -21,7 +21,7 @@ private:
 	size_t IndexOfFullMoveCounter;//before halfmove clock
 public:
 	Board(const std::string& FenString);
-	std::array<unsigned int, 64> GetPositionFromFEN();
+	std::array<uint8_t, 64> GetPositionFromFEN();
 	uint32_t MoveNum();
 	canCastle GetCanCastle();
 	uint32_t GetPawnMoveSquare();
